refactor: Phase 3 slice 3c — downloads/ffmpeg/cleanup/downloaded-anime routers (#84)
Third slice (3c) of Phase 3 — the biggest by handler count.
Migrates 26 handlers (downloads, ffmpeg, cleanup, downloaded-anime)
behind the router pattern. Behavior-identical, no IPC surface change.

* New src/main/ipc/downloads.ipc.ts (17 handlers) — full
  DownloadManager surface: DOWNLOAD_{ENQUEUE, PAUSE, RESUME, RESTART,
  RESTART_ALL_FAILED, PAUSE_ALL, RESUME_ALL, CANCEL,
  CANCEL_BY_EPISODE, GET_QUEUE, CANCEL_MERGE, CLEAR_COMPLETED, MERGE,
  SCAN_MERGE, FIX_METADATA, PICK_DIR} + DOWNLOADED_EPISODES_GET. The
  cancel-by-episode handler still prunes orphaned metadata and removes
  the show from downloadedAnime when no files survive. SCAN_MERGE
  still broadcasts EVENT_CHANNELS.SCAN_MERGE_PROGRESS. FIX_METADATA
  still uses fluent-ffmpeg to rewrite subtitle disposition/metadata
  on existing MKVs and broadcasts FIX_METADATA_PROGRESS.

* New src/main/ipc/ffmpeg.ipc.ts (2 handlers) — FFMPEG_CHECK runs
  the injected checkFfmpeg; FFMPEG_DELETE removes the binaries,
  clears the ffbinaries zip cache, and resets the mutable
  ffmpegPath/ffprobePath module variables via clearFfmpegPaths.

* New src/main/ipc/cleanup.ipc.ts (5 handlers) — CLEANUP_GET_SIZE
  (via the injected sumShowFiles), CLEANUP_GET_ACTIVE_DOWNLOADS,
  CLEANUP_EXECUTE (deletes across coldStorageService.dirsForScan(),
  prunes downloadedEpisodes/anime-cache/skip-detection),
  CLEANUP_{GET,SET}_SNOOZED.

* New src/main/ipc/downloaded-anime.ipc.ts (2 handlers) —
  DOWNLOADED_ANIME_ADD/DELETE; delete still cascades to all
  storage dirs, anime cache, and skip-detection cache+detections.

* AppDeps grows substantially to support the new routers:
  + downloadManager, coldStorageService, skipAnalysisService
  + live ffmpeg-path accessors: getFfmpegPath, getFfprobePath,
    clearFfmpegPaths (FFMPEG_DELETE mutates them, so handlers must
    read through getters every call instead of capturing at register)
  + invalidateFileCache (closes over the in-memory file-check Map)
  + sumShowFiles, checkFfmpeg, getFfmpegDir, getDisplayName closures
  + FfmpegInfo type definition co-located with AppDeps

* src/main/index.ts shrinks by ~485 LOC of removed handlers and
  the now-unused DownloadRequest type import. registerIpcHandlers()
  now passes the extended dep bundle and keeps 71 handlers
  (downloaded to skip-detector/storage/syncplay/player/etc.).

* DESIGN.md adds 4 new module rows.

Stacked on phase3-pr2-anime-library-home; refs #102, #84.
