+++ DESIGN.md
@@ -36,6 +36,10 @@ Renderer (Vue)  --ipcRenderer.invoke-->  Preload (bridge)  --ipcMain.handle-->  
 | `src/main/ipc/cache.ipc.ts` | Anime-detail cache router: `CHANNELS.{GET_ANIME_CACHE (24h TTL), SET_ANIME_CACHE (gated by library ∪ downloadedAnime), CACHE_GET_POSTER}` |
 | `src/main/ipc/home.ipc.ts` | `CHANNELS.HOME_GET_CONTINUE_WATCHING` — thin shell wrapping the pure helpers in `lib/continue-watching.ts`, plus the bounded-concurrency smotret fetch loop for rows that are still nameless / posterless |
 | `src/main/lib/continue-watching.ts` | Pure helpers `buildContinueWatchingEntries(inputs)` + `finalizeContinueWatchingEntries(draft, fetched)` — collapse `watchProgress` + `shikimoriUserRates` into resume + next-episode rows, apply the Shikimori `rate.updated_at` override, hide completed shows, generate the `unresolvedIds` fetch list, finalize ordering and the 24-row cap. Lifted out of `HOME_GET_CONTINUE_WATCHING` (Phase 2 deferral) for unit-testable logic |
+| `src/main/ipc/downloads.ipc.ts` | Downloads router (Phase 3 slice 3c): `CHANNELS.{DOWNLOAD_ENQUEUE, DOWNLOAD_{PAUSE, RESUME, RESTART, RESTART_ALL_FAILED, PAUSE_ALL, RESUME_ALL, CANCEL, CANCEL_BY_EPISODE, GET_QUEUE, CANCEL_MERGE, CLEAR_COMPLETED, MERGE, SCAN_MERGE, FIX_METADATA, PICK_DIR}, DOWNLOADED_EPISODES_GET}` — wraps `DownloadManager`, the `coldStorageService.pruneDownloadedEpisode` prune-on-cancel, and the `fluent-ffmpeg`-driven fix-metadata pass. Reads ffmpeg/ffprobe paths through `getFfmpegPath()`/`getFfprobePath()` live getters so `FFMPEG_DELETE` clearing the paths is observed without re-registration |
+| `src/main/ipc/ffmpeg.ipc.ts` | Ffmpeg router: `CHANNELS.{FFMPEG_CHECK, FFMPEG_DELETE}` — checks via the injected `checkFfmpeg`, deletes binaries + `ffbinaries-cache` zip cache, and clears the mutable `ffmpegPath`/`ffprobePath` module variables via `clearFfmpegPaths()` |
+| `src/main/ipc/cleanup.ipc.ts` | Cleanup router: `CHANNELS.{CLEANUP_GET_SIZE, CLEANUP_GET_ACTIVE_DOWNLOADS, CLEANUP_EXECUTE, CLEANUP_GET_SNOOZED, CLEANUP_SET_SNOOZED}` — uses `coldStorageService.dirsForScan()` to delete the show across hot+cold dirs, prunes `downloadedEpisodes` keys, drops anime/skip-detection cache entries, snoozes per anime id |
+| `src/main/ipc/downloaded-anime.ipc.ts` | Downloaded-anime router: `CHANNELS.{DOWNLOADED_ANIME_ADD, DOWNLOADED_ANIME_DELETE}` — delete cascades to all storage dirs, anime-cache entry, skip-detection cache + detections |
 | `src/main/smotret-api.ts` | Smotret-Anime API client: search, anime/episode details, embed, subtitles, token validation |
 | `src/main/download-manager.ts` | Download queue, concurrent downloads, progress, ffmpeg merge, scan-merge |
 | `src/main/shikimori.ts` | Shikimori API client: OAuth, token refresh, user rates, anime list |
